Lots of change. See README for full list. After security is complete and the code is commented, will merge with master.

+++ app/controllers/Crags.java
@@ -20,8 +20,11 @@ public class Crags extends Controller {
     }
 
     public static Result deleteCrag(String cragname){
-        Crag.find.ref(cragname).delete();
-        return ok();
+        if (request().username().equals("ryanbrushett")){
+            Crag.find.ref(cragname).delete();
+            return ok();
+        }
+        return unauthorized("You are not authorized to delete locations");
     }
 
     public static Result listCrags(){
